ReorganizeModuleBackends: Remove module_backend_config
All of this configuration is now handled by the LocalModelFinder when it
is initialized.

https://github.com/caikit/caikit/issues/54
